refactor tests in 3 categories unit, e2e and electron
