+++ .github/workflows/node-ci.js.yml
@@ -23,7 +23,6 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
       - run: |
-          node -v
           Xvfb :99 -screen 0 1024x768x16 &
           export DISPLAY=:99
           npm install
@@ -32,3 +31,4 @@ jobs:
           npm run test:css
           npm run test:unit
           npm run test:e2e
+          npm run test:electron
