+++ tests/e2e/ipWhitelist_spec.js
@@ -1,23 +1,13 @@
-const helpers = require("./global-setup");
 const fetch = require("node-fetch");
+const app = require("../../js/app.js");
 
 describe("ipWhitelist directive configuration", function () {
-	helpers.setupTimeout(this);
-
-	let app = null;
-
-	beforeEach(function () {
-		return helpers
-			.startApplication({
-				args: ["js/electron.js"]
-			})
-			.then(function (startedApp) {
-				app = startedApp;
-			});
+	beforeAll(function () {
+		app.start();
 	});
 
-	afterEach(function () {
-		return helpers.stopApplication(app);
+	afterAll(function () {
+		app.stop();
 	});
 
 	describe("Set ipWhitelist without access", function () {
