+++ tests/e2e/vendor_spec.js
@@ -1,24 +1,15 @@
-const helpers = require("./global-setup");
 const fetch = require("node-fetch");
+const app = require("../../js/app.js");
 
 describe("Vendors", function () {
-	helpers.setupTimeout(this);
-
-	let app = null;
-
 	beforeAll(function () {
 		process.env.MM_CONFIG_FILE = "tests/configs/env.js";
-		return helpers
-			.startApplication({
-				args: ["js/electron.js"]
-			})
-			.then(function (startedApp) {
-				app = startedApp;
-			});
+
+		app.start();
 	});
 
 	afterAll(function () {
-		return helpers.stopApplication(app);
+		app.stop();
 	});
 
 	describe("Get list vendors", function () {
