+++ tests/electron/global-setup.js
