+++ tests/electron/modules/basic-auth.js
