+++ tests/electron/modules/clock_spec.js
