+++ tests/electron/modules/compliments_spec.js
