+++ tests/electron/modules/mocks/index.js
