+++ tests/electron/modules/mocks/weather_current.js
