+++ tests/electron/modules/mocks/weather_forecast.js
