+++ tests/electron/modules/newsfeed_spec.js
