+++ tests/electron/modules_display_spec.js
