+++ tests/electron/translations_spec.js
